requirements.txt
python-telegram-bot==20.0
google-api-python-client
google-auth-oauthlib
google-auth
Pillow
python-dotenv
requests
